requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
face-recognition==1.3.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
python-multipart==0.0.9
numpy==1.26.4
setuptools<81
aiofiles==24.1.0
onnxruntime>=1.19.0
